backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
aiofiles==24.1.0
Pillow==11.1.0
